+++ calibre-plugin/__init__.py
@@ -21,7 +21,7 @@ if False:
 
 load_translations()
 
-__version__ = (0, 1, 9, 2)
+__version__ = (0, 1, 9, 3)
 PLUGIN_NAME = "overdrive_libby"
 PLUGIN_ICON = "images/plugin.svg"
 PLUGINS_FOLDER_NAME = "plugins"